Fix : 서류,면접 평가기준 합에서 평균으로

+++ src/main/java/com/cluting/clutingbackend/evaluation/dto/interview/InterviewEvaluationResponseDto.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @AllArgsConstructor
 public class InterviewEvaluationResponseDto {
     private Long interviewId;
-    private Integer totalScore;
+    private Double totalScore;
     private String comment;
     private String status;
 }
